changed

+++ method.go
@@ -3568,7 +3568,7 @@ func (client *Client) CreatePrivateChat(userID int64, force bool) (*Chat, error)
 // CreateBasicGroupChat Returns an existing chat corresponding to a known basic group
 // @param basicGroupID Basic group identifier
 // @param force If true, the chat will be created without network request. In this case all information about the chat except its type, title and photo can be incorrect
-func (client *Client) CreateBasicGroupChat(basicGroupID int32, force bool) (*Chat, error) {
+func (client *Client) CreateBasicGroupChat(basicGroupID int64, force bool) (*Chat, error) {
 	result, err := client.SendAndCatch(UpdateData{
 		"@type":          "createBasicGroupChat",
 		"basic_group_id": basicGroupID,
@@ -3592,7 +3592,7 @@ func (client *Client) CreateBasicGroupChat(basicGroupID int32, force bool) (*Cha
 // CreateSupergroupChat Returns an existing chat corresponding to a known supergroup or channel
 // @param supergroupID Supergroup or channel identifier
 // @param force If true, the chat will be created without network request. In this case all information about the chat except its type, title and photo can be incorrect
-func (client *Client) CreateSupergroupChat(supergroupID int32, force bool) (*Chat, error) {
+func (client *Client) CreateSupergroupChat(supergroupID int64, force bool) (*Chat, error) {
 	result, err := client.SendAndCatch(UpdateData{
 		"@type":         "createSupergroupChat",
 		"supergroup_id": supergroupID,
@@ -6905,7 +6905,7 @@ func (client *Client) DisconnectAllWebsites() (*Ok, error) {
 // SetSupergroupUsername Changes the username of a supergroup or channel, requires owner privileges in the supergroup or channel
 // @param supergroupID Identifier of the supergroup or channel
 // @param username New value of the username. Use an empty string to remove the username
-func (client *Client) SetSupergroupUsername(supergroupID int32, username string) (*Ok, error) {
+func (client *Client) SetSupergroupUsername(supergroupID int64, username string) (*Ok, error) {
 	result, err := client.SendAndCatch(UpdateData{
 		"@type":         "setSupergroupUsername",
 		"supergroup_id": supergroupID,
@@ -6929,7 +6929,7 @@ func (client *Client) SetSupergroupUsername(supergroupID int32, username string)
 // SetSupergroupStickerSet Changes the sticker set of a supergroup; requires can_change_info administrator right
 // @param supergroupID Identifier of the supergroup
 // @param stickerSetID New value of the supergroup sticker set identifier. Use 0 to remove the supergroup sticker set
-func (client *Client) SetSupergroupStickerSet(supergroupID int32, stickerSetID JSONInt64) (*Ok, error) {
+func (client *Client) SetSupergroupStickerSet(supergroupID int64, stickerSetID JSONInt64) (*Ok, error) {
 	result, err := client.SendAndCatch(UpdateData{
 		"@type":          "setSupergroupStickerSet",
 		"supergroup_id":  supergroupID,
@@ -6953,7 +6953,7 @@ func (client *Client) SetSupergroupStickerSet(supergroupID int32, stickerSetID J
 // ToggleSupergroupSignMessages Toggles sender signatures messages sent in a channel; requires can_change_info administrator right
 // @param supergroupID Identifier of the channel
 // @param signMessages New value of sign_messages
-func (client *Client) ToggleSupergroupSignMessages(supergroupID int32, signMessages bool) (*Ok, error) {
+func (client *Client) ToggleSupergroupSignMessages(supergroupID int64, signMessages bool) (*Ok, error) {
 	result, err := client.SendAndCatch(UpdateData{
 		"@type":         "toggleSupergroupSignMessages",
 		"supergroup_id": supergroupID,
@@ -6977,7 +6977,7 @@ func (client *Client) ToggleSupergroupSignMessages(supergroupID int32, signMessa
 // ToggleSupergroupIsAllHistoryAvailable Toggles whether the message history of a supergroup is available to new members; requires can_change_info administrator right
 // @param supergroupID The identifier of the supergroup
 // @param isAllHistoryAvailable The new value of is_all_history_available
-func (client *Client) ToggleSupergroupIsAllHistoryAvailable(supergroupID int32, isAllHistoryAvailable bool) (*Ok, error) {
+func (client *Client) ToggleSupergroupIsAllHistoryAvailable(supergroupID int64, isAllHistoryAvailable bool) (*Ok, error) {
 	result, err := client.SendAndCatch(UpdateData{
 		"@type":                    "toggleSupergroupIsAllHistoryAvailable",
 		"supergroup_id":            supergroupID,
@@ -7029,7 +7029,7 @@ func (client *Client) ReportSupergroupSpam(supergroupID int64, userID int64, mes
 // @param filter The type of users to return. By default, supergroupMembersFilterRecent
 // @param offset Number of users to skip
 // @param limit The maximum number of users be returned; up to 200
-func (client *Client) GetSupergroupMembers(supergroupID int32, filter SupergroupMembersFilter, offset int32, limit int32) (*ChatMembers, error) {
+func (client *Client) GetSupergroupMembers(supergroupID int64, filter SupergroupMembersFilter, offset int32, limit int32) (*ChatMembers, error) {
 	result, err := client.SendAndCatch(UpdateData{
 		"@type":         "getSupergroupMembers",
 		"supergroup_id": supergroupID,
